refactor: move authentication-related code to it's own folder
