+++ src/lib/server/authentication/basic.ts
@@ -1,4 +1,4 @@
-export const parseAuthorizationHeader = (header: string) => {
+export const parseBasicAuthorizationHeader = (header: string) => {
 	const credentials = header.match(/Basic (.+)/)?.[1];
 	const split = atob(credentials ?? '').split(':');
 	return { username: split.at(0)!, password: split.at(1) ?? '' };
@@ -15,7 +15,7 @@ if (import.meta.vitest) {
 		{ header: 'Basic eWFzc2VyOmFo', username: 'yasser', password: 'ah' }
 	])('`$header` parsed into `$username:$password`', ({ header, password, username }) => {
 		it('should parse', () => {
-			expect(parseAuthorizationHeader(header)).toStrictEqual({ username, password });
+			expect(parseBasicAuthorizationHeader(header)).toStrictEqual({ username, password });
 		});
 	});
 }
